backend/requirements.txt
Flask==2.3.2
openai==0.27.8
python-dotenv==1.0.0
langchain-anthropic
anthropic
pydantic